# server/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Consolidation:
    properties:
      createdAt:
        type: string
      id:
        type: integer
    type: object
...
  model.Room:
    properties:
      attendees:
        items:
          $ref: '#/definitions/model.User'
        type: array
      consolidated:
        type: string
      createdAt:
        type: string
      date:
        type: string
      description:
        type: string
      host:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Associations
      hostId:
        type: integer
      id:
        type: string
      imageUrl:
        type: string
      isClosed:
        type: boolean
      isPrivate:
        type: boolean
      name:
        type: string
      noOfAttendees:
        type: integer
      time:
        type: string
      updatedAt:
        type: string
      venue:
        type: string
      venuePlaceId:
        type: string
      venueUrl:
        type: string
    type: object
...
  model.Subscription:
    properties:
      auth:
        type: string
      endpoint:
        type: string
      id:
        type: string
      p256dh:
        type: string
      userId:
        type: integer
    type: object
  model.Transaction:
    properties:
      amount:
        type: number
      consolidation:
        $ref: '#/definitions/model.Consolidation'
      consolidationId:
        type: integer
      id:
        type: integer
      isPaid:
        type: boolean
      paidOn:
        type: string
      payee:
        $ref: '#/definitions/model.User'
      payeeId:
        type: integer
      payer:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Associations
      payerId:
        type: integer
    type: object
  model.User:
    properties:
      email:
        type: string
      friends:
        items:
          $ref: '#/definitions/model.User'
        type: array
      id:
        type: integer
      isEmailValid:
        type: boolean
      isOnline:
        type: boolean
      lastSeen:
        type: string
      noOfFriends:
        type: integer
      noOfPendingFriendRequests:
        type: integer
      noOfRoomInvites:
        type: integer
      noOfRooms:
        type: integer
      password:
        type: string
      pictureUrl:
        type: string
      registeredAt:
        type: string
      rooms:
        description: Associations
        items:
          $ref: '#/definitions/model.Room'
        type: array
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  utils.EmptyApiResponse:
    properties:
      data:
        type: object
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for JustJio.
  title: JustJio API
  version: "1.0"
paths:
...
  /subscriptions:
    post:
      consumes:
      - application/json
      description: Creates a new push notification subscription for the user
      parameters:
      - description: Subscription details
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: Subscription created successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/model.Subscription'
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Create subscription
      tags:
      - Subscriptions
  /subscriptions/{endpoint}:
    get:
      consumes:
      - application/json
      description: Retrieves a push notification subscription by its endpoint URL
      parameters:
      - description: URL-encoded subscription endpoint
        in: path
        name: endpoint
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subscription retrieved successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/model.Subscription'
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Invalid endpoint URL
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Subscription not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Get subscription by endpoint
      tags:
      - Subscriptions
  /subscriptions/{subId}:
    delete:
      consumes:
      - application/json
      description: Deletes a push notification subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: subId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subscription deleted successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Subscription not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Delete subscription
      tags:
      - Subscriptions
  /transactions:
    get:
      consumes:
      - application/json
      description: Retrieves transactions for the authenticated user, optionally filtered
        by payment status
      parameters:
      - default: false
        description: Filter by payment status
        in: query
        name: isPaid
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved transactions successfully
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.Transaction'
                type: array
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: No transactions found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Get user transactions
      tags:
      - Transactions
  /transactions/{txId}/settle:
    patch:
      consumes:
      - application/json
      description: Marks a transaction as paid by the authenticated user
      parameters:
      - description: Transaction ID
        in: path
        name: txId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paid transactions successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "400":
          description: Invalid payer or bad request
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Transaction not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "409":
          description: Transaction already settled
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Settle transaction
      tags:
      - Transactions
...
schemes:
- http
- https
swagger: "2.0"