# server/api/handlers/subscription.go
package handlers

import (
	"errors"
	"net/url"

	log "github.com/sirupsen/logrus"

	"github.com/RowenTey/JustJio/server/api/model"
	"github.com/RowenTey/JustJio/server/api/services"
	"github.com/RowenTey/JustJio/server/api/utils"
	"github.com/gofiber/fiber/v2"
)

type SubscriptionHandler struct {
	subscriptionService *services.SubscriptionService
	logger              *log.Entry
}

func NewSubscriptionHandler(
	subscriptionService *services.SubscriptionService,
	logger *log.Logger,
) *SubscriptionHandler {
	return &SubscriptionHandler{
		subscriptionService: subscriptionService,
		logger:              utils.AddServiceField(logger, "SubscriptionHandler"),
	}
}

// CreateSubscription creates a new push notification subscription
// @Summary Create subscription
// @Description Creates a new push notification subscription for the user
// @Tags Subscriptions
// @Accept json
// @Produce json
// @Param subscription body model.Subscription true "Subscription details"
// @Success 200 {object} object{status=string,message=string,data=model.Subscription} "Subscription created successfully"
// @Failure 400 {object} utils.EmptyApiResponse "Invalid input or missing required fields"
// @Failure 500 {object} utils.EmptyApiResponse "Internal server error"
// @Router /subscriptions [post]
func (h *SubscriptionHandler) CreateSubscription(c *fiber.Ctx) error {
	var subscription model.Subscription
	if err := c.BodyParser(&subscription); err != nil {
		return utils.HandleInvalidInputError(c, err)
	}

	if subscription.Auth == "" || subscription.P256dh == "" || subscription.Endpoint == "" {
		return utils.HandleInvalidInputError(c, errors.New("missing required fields"))
	}

	createdSubscription, err := h.subscriptionService.CreateSubscription(&subscription)
	if err != nil {
		return utils.HandleInternalServerError(c, err)
	}

	h.logger.Info("Subscription created successfully: ", createdSubscription.ID)
	return utils.HandleSuccess(c, "Subscription created successfully", createdSubscription)
}

// GetSubscriptionByEndpoint retrieves a subscription by endpoint
// @Summary Get subscription by endpoint
// @Description Retrieves a push notification subscription by its endpoint URL
// @Tags Subscriptions
// @Accept json
// @Produce json
// @Param endpoint path string true "URL-encoded subscription endpoint"
// @Success 200 {object} object{status=string,message=string,data=model.Subscription} "Subscription retrieved successfully"
// @Failure 400 {object} utils.EmptyApiResponse "Invalid endpoint URL"
// @Failure 404 {object} utils.EmptyApiResponse "Subscription not found"
// @Failure 500 {object} utils.EmptyApiResponse "Internal server error"
// @Router /subscriptions/{endpoint} [get]
func (h *SubscriptionHandler) GetSubscriptionByEndpoint(c *fiber.Ctx) error {
	endpoint := c.Params("endpoint")
	decodedEndpoint, err := url.QueryUnescape(endpoint)
	if err != nil {
		return utils.HandleInvalidInputError(c, err)
	}

	subscription, err := h.subscriptionService.GetSubscriptionsByEndpoint(decodedEndpoint)
	if err != nil {
		return utils.HandleNotFoundOrInternalError(c, err, "Subscription not found")
	}

	return utils.HandleSuccess(c, "Subscription retrieved successfully", subscription)
}

// DeleteSubscription deletes a subscription
// @Summary Delete subscription
// @Description Deletes a push notification subscription by ID
// @Tags Subscriptions
// @Accept json
// @Produce json
// @Param subId path string true "Subscription ID"
// @Success 200 {object} utils.EmptyApiResponse "Subscription deleted successfully"
// @Failure 404 {object} utils.EmptyApiResponse "Subscription not found"
// @Failure 500 {object} utils.EmptyApiResponse "Internal server error"
// @Router /subscriptions/{subId} [delete]
func (h *SubscriptionHandler) DeleteSubscription(c *fiber.Ctx) error {
	subId := c.Params("subId")

	if err := h.subscriptionService.DeleteSubscription(subId); err != nil {
		return utils.HandleNotFoundOrInternalError(c, err, "Subscription not found")
	}

	return utils.HandleSuccess[any](c, "Subscription deleted successfully", nil)
}